requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==3.0.1
WTForms==3.1.1
